pyproject.toml
[tool.poetry]
name = "badge-generator-be"
version = "0.2.0"
description = "Badge Generator for Annual Conference"
authors = [
    "Kevin Lin <kevin.cj.lin@toastmasters.org.tw>",
    "Sherry Lee <Sherry.Lee@toastmasters.org.tw>"
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.28.0"}
pillow = "^10.3.0"
python-multipart = "^0.0.9"
reportlab = "^4.1.0"
poetry = "^1.8.3"
pydantic-settings = "^2.2.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.3.7"

[tool.poe.tasks]
format = "ruff format"
lint = "ruff check --fix"
run = "uvicorn main:app --reload"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
